Crude

+++ src/main/java/br/com/etec/cinehas/entity/Sala.java
@@ -11,9 +11,12 @@ public class Sala {
     @Id // Esse campo é a chave primária na tabela
     @Column(name = "ID_SALA")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @Column(name = "TX_NOME")
     private String nome;
     @Column(name = "NR_PRECO")
     private double preco;
+
+
+
 }
